pyproject.toml
[tool.poetry]
name = "sepdpc"
version = "0.1.0"
description = ""
authors = ["Arne Ottens <arne.ottens@starburst.io>"]
readme = "README.md"
packages = [{include = "sepdpc", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.9.0"
pydantic = "^2.1.1"
pyyaml = "^6.0.1"
deepdiff = "^6.3.1"
python-dotenv = "^1.0.0"
adsepra = {path = "/Users/arne.ottens/devel/PycharmProjects/adsepra"}
rich = "^13.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
